// src/app/core/api/generated/model/dayScheduleDto.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DayType } from './dayType';


export interface DayScheduleDto { 
    dayNumber?: number;
    shiftId?: string | null;
    dayType?: DayType;
    isHoliday?: boolean;
    notes?: string | null;
}